requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-dotenv
python-multipart
psycopg2-binary
bcrypt
passlib
email-validator
